refactor: (#741) Response를 Omit하여 만든 타입의 가독성을 높이기 위해 Response 타입을 이동 같은 파일에 있어야 코드 가독성이 높아질 것으로 생각했습니다
