+++ frontend/src/types/notice.ts
@@ -1,6 +1,22 @@
-import { NoticeListResponse, NoticeResponse } from '@api/notice';
+import { StringDate, StringDateOnly } from './time';
 
-import { StringDateOnly } from './time';
+export interface NoticeResponse {
+  id: number;
+  title: string;
+  content: string;
+  createdAt: StringDate;
+  deadline: StringDate;
+  bannerTitle: string;
+  bannerSubtitle: string;
+}
+
+export interface NoticeListResponse {
+  totalPageNumber: number;
+  currentPageNumber: number;
+  notices: NoticeResponse[];
+}
+
+export type NoticeRequest = Omit<Notice, 'createdAt' | 'id'>;
 
 export type Notice = Omit<NoticeResponse, 'createdAt'> & { createdAt: StringDateOnly };
